十年OI一场空，不开longlong见祖宗

+++ src/entity/student.h
@@ -15,7 +15,7 @@ typedef struct _student_t {
     int id;
     char name[50];
     gender_t gender;
-    long phone;
+    long long phone;  // 十年OI一场空，不开longlong见祖宗
     char password[128];
     char salt[16];
 } student_t;
